+++ test/HStargate.test.js
@@ -132,8 +132,8 @@ contract('Stargate', function ([_, user, user2]) {
     const payload = '0x';
 
     describe('Native', function () {
-      if (chainId == 137 || chainId == 43114) {
-        // Stargate does not support MATIC / AVAX for now
+      if (chainId == 137 || chainId == 250 || chainId == 43114) {
+        // Stargate does not support MATIC / FTM / AVAX for now
         return;
       }
 
@@ -156,18 +156,18 @@ contract('Stargate', function ([_, user, user2]) {
           { dstGasForCall: 0, dstNativeAmount: 0, dstNativeAddr: '0x' } // lzTxObj
         );
         const fee = fees[0];
-        const value = amountIn.add(fee);
         const data = abi.simpleEncode(
-          'swapETH(uint256,uint16,address,uint256,uint256,address)',
-          value,
+          'swapETH(uint16,address,uint256,uint256,uint256,address)',
           dstChainId,
           refundAddress,
+          amountIn,
           fee,
           amountOutMin,
           receiver
         );
 
         // Execute
+        const value = amountIn.add(fee);
         const receipt = await this.proxy.execMock(to, data, {
           from: user,
           value: value,
@@ -201,18 +201,18 @@ contract('Stargate', function ([_, user, user2]) {
           { dstGasForCall: 0, dstNativeAmount: 0, dstNativeAddr: '0x' } // lzTxObj
         );
         const fee = fees[0];
-        const value = amountIn.add(fee);
         const data = abi.simpleEncode(
-          'swapETH(uint256,uint16,address,uint256,uint256,address)',
-          MAX_UINT256,
+          'swapETH(uint16,address,uint256,uint256,uint256,address)',
           dstChainId,
           refundAddress,
+          MAX_UINT256,
           fee,
           amountOutMin,
           receiver
         );
 
         // Execute
+        const value = amountIn.add(fee);
         const receipt = await this.proxy.execMock(to, data, {
           from: user,
           value: value,
@@ -247,18 +247,18 @@ contract('Stargate', function ([_, user, user2]) {
           { dstGasForCall: 0, dstNativeAmount: 0, dstNativeAddr: '0x' } // lzTxObj
         );
         const fee = fees[0];
-        const value = amountIn.add(fee);
         const data = abi.simpleEncode(
-          'swapETH(uint256,uint16,address,uint256,uint256,address)',
-          value,
+          'swapETH(uint16,address,uint256,uint256,uint256,address)',
           dstChainId,
           refundAddress,
+          amountIn,
           fee,
           amountOutMin,
           receiver
         );
 
         // Execute
+        const value = amountIn.add(fee);
         const receipt = await this.proxy.execMock(to, data, {
           from: user,
           value: value,
@@ -294,18 +294,18 @@ contract('Stargate', function ([_, user, user2]) {
         const fee = fees[0];
         const extraFee = fee.mul(new BN('2'));
         const totalFee = fee.add(extraFee);
-        const value = amountIn.add(totalFee);
         const data = abi.simpleEncode(
-          'swapETH(uint256,uint16,address,uint256,uint256,address)',
-          value,
+          'swapETH(uint16,address,uint256,uint256,uint256,address)',
           dstChainId,
           refundAddress,
+          amountIn,
           totalFee,
           amountOutMin,
           receiver
         );
 
         // Execute
+        const value = amountIn.add(totalFee);
         const receipt = await this.proxy.execMock(to, data, {
           from: user,
           value: value,
@@ -341,18 +341,18 @@ contract('Stargate', function ([_, user, user2]) {
           { dstGasForCall: 0, dstNativeAmount: 0, dstNativeAddr: '0x' } // lzTxObj
         );
         const fee = fees[0];
-        const value = amountIn.add(fee);
         const data = abi.simpleEncode(
-          'swapETH(uint256,uint16,address,uint256,uint256,address)',
-          value,
+          'swapETH(uint16,address,uint256,uint256,uint256,address)',
           dstChainId,
           refundAddress,
+          amountIn,
           fee,
           amountOutMin,
           receiver
         );
 
         // Execute
+        const value = amountIn.add(fee);
         await expectRevert(
           this.proxy.execMock(to, data, {
             from: user,
@@ -368,18 +368,18 @@ contract('Stargate', function ([_, user, user2]) {
         const refundAddress = this.proxy.address;
         const to = this.hStargate.address;
         const fee = ether('1'); // Use fixed fee because of unknown chain id
-        const value = amountIn.add(fee);
         const data = abi.simpleEncode(
-          'swapETH(uint256,uint16,address,uint256,uint256,address)',
-          value,
+          'swapETH(uint16,address,uint256,uint256,uint256,address)',
           dstChainId,
           refundAddress,
+          amountIn,
           fee,
           amountOutMin,
           receiver
         );
 
         // Execute
+        const value = amountIn.add(fee);
         await expectRevert(
           this.proxy.execMock(to, data, {
             from: user,
@@ -426,18 +426,18 @@ contract('Stargate', function ([_, user, user2]) {
         const refundAddress = this.proxy.address;
         const to = this.hStargate.address;
         const fee = ether('0');
-        const value = amountIn.add(fee);
         const data = abi.simpleEncode(
-          'swapETH(uint256,uint16,address,uint256,uint256,address)',
-          value,
+          'swapETH(uint16,address,uint256,uint256,uint256,address)',
           dstChainId,
           refundAddress,
+          amountIn,
           fee,
           amountOutMin,
           receiver
         );
 
         // Execute
+        const value = amountIn.add(fee);
         await expectRevert(
           this.proxy.execMock(to, data, {
             from: user,
@@ -459,18 +459,18 @@ contract('Stargate', function ([_, user, user2]) {
           { dstGasForCall: 0, dstNativeAmount: 0, dstNativeAddr: '0x' } // lzTxObj
         );
         const fee = fees[0].div(new BN('2'));
-        const value = amountIn.add(fee);
         const data = abi.simpleEncode(
-          'swapETH(uint256,uint16,address,uint256,uint256,address)',
-          value,
+          'swapETH(uint16,address,uint256,uint256,uint256,address)',
           dstChainId,
           refundAddress,
+          amountIn,
           fee,
           amountOutMin,
           receiver
         );
 
         // Execute
+        const value = amountIn.add(fee);
         await expectRevert(
           this.proxy.execMock(to, data, {
             from: user,
@@ -494,18 +494,18 @@ contract('Stargate', function ([_, user, user2]) {
           { dstGasForCall: 0, dstNativeAmount: 0, dstNativeAddr: '0x' } // lzTxObj
         );
         const fee = fees[0];
-        const value = amountIn.add(fee);
         const data = abi.simpleEncode(
-          'swapETH(uint256,uint16,address,uint256,uint256,address)',
-          value,
+          'swapETH(uint16,address,uint256,uint256,uint256,address)',
           dstChainId,
           refundAddress,
+          amountIn,
           fee,
           amountOutMin,
           receiver
         );
 
         // Execute
+        const value = amountIn.add(fee);
         await expectRevert(
           this.proxy.execMock(to, data, {
             from: user,
@@ -528,18 +528,18 @@ contract('Stargate', function ([_, user, user2]) {
           { dstGasForCall: 0, dstNativeAmount: 0, dstNativeAddr: '0x' } // lzTxObj
         );
         const fee = fees[0];
-        const value = amountIn.add(fee);
         const data = abi.simpleEncode(
-          'swapETH(uint256,uint16,address,uint256,uint256,address)',
-          value,
+          'swapETH(uint16,address,uint256,uint256,uint256,address)',
           dstChainId,
           refundAddress,
+          amountIn,
           fee,
           amountOutMin,
           receiver
         );
 
         // Execute
+        const value = amountIn.add(fee);
         await expectRevert(
           this.proxy.execMock(to, data, {
             from: user,
@@ -566,18 +566,18 @@ contract('Stargate', function ([_, user, user2]) {
           { dstGasForCall: 0, dstNativeAmount: 0, dstNativeAddr: '0x' } // lzTxObj
         );
         const fee = fees[0];
-        const value = amountIn.add(fee);
         const data = abi.simpleEncode(
-          'swapETH(uint256,uint16,address,uint256,uint256,address)',
-          value,
+          'swapETH(uint16,address,uint256,uint256,uint256,address)',
           dstChainId,
           refundAddress,
+          amountIn,
           fee,
           amountOutMin,
           receiver
         );
 
         // Execute
+        const value = amountIn.add(fee);
         await expectRevert(
           this.proxy.execMock(to, data, {
             from: user,
@@ -600,18 +600,18 @@ contract('Stargate', function ([_, user, user2]) {
           { dstGasForCall: 0, dstNativeAmount: 0, dstNativeAddr: '0x' } // lzTxObj
         );
         const fee = fees[0];
-        const value = amountIn.add(fee);
         const data = abi.simpleEncode(
-          'swapETH(uint256,uint16,address,uint256,uint256,address)',
-          value,
+          'swapETH(uint16,address,uint256,uint256,uint256,address)',
           dstChainId,
           refundAddress,
+          amountIn,
           fee,
           amountOutMin,
           receiver
         );
 
         // Execute
+        const value = amountIn.add(fee);
         await expectRevert(
           this.proxy.execMock(to, data, {
             from: user,
@@ -634,18 +634,18 @@ contract('Stargate', function ([_, user, user2]) {
           { dstGasForCall: 0, dstNativeAmount: 0, dstNativeAddr: '0x' } // lzTxObj
         );
         const fee = fees[0];
-        const value = amountIn.add(fee);
         const data = abi.simpleEncode(
-          'swapETH(uint256,uint16,address,uint256,uint256,address)',
-          value,
+          'swapETH(uint16,address,uint256,uint256,uint256,address)',
           dstChainId,
           refundAddress,
+          amountIn,
           fee,
           amountOutMin,
           receiver
         );
 
         // Execute
+        const value = amountIn.add(fee);
         await expectRevert(
           this.proxy.execMock(to, data, {
             from: user,
@@ -1261,22 +1261,18 @@ contract('Stargate', function ([_, user, user2]) {
     });
 
     describe('STG Token', function () {
-      const version = new BN('1');
+      const ptSend = 0;
       const inputTokenAddr = STG_TOKEN;
       const amountIn = ether('1');
-      const dstGas = new BN('85000'); // stg gas value from a real tx
-      const receiverBytes = abi.solidityPack(['address'], [receiver]);
-      const payload = abi.rawEncode(
-        ['bytes', 'uint256'],
-        [receiverBytes, amountIn]
+      const ld2sdRate = mwei('10000'); // 10^10 (private param)
+      const receiverBytes32 = abi.rawEncode(['address'], [receiver]);
+      const payload = abi.solidityPack(
+        ['uint8', 'bytes32', 'uint64'],
+        [ptSend, receiverBytes32, amountIn.div(ld2sdRate)]
       );
-      const adapterParam = abi.solidityPack(
-        ['uint16', 'uint256'],
-        [version, dstGas]
-      );
+      const adapterParam = abi.solidityPack(['string'], ['']);
 
       let inputTokenPoolBefore;
-      let isMain;
 
       before(async function () {
         inputTokenProvider = STARGATE_MULTISIG;
@@ -1295,11 +1291,11 @@ contract('Stargate', function ([_, user, user2]) {
         await this.proxy.updateTokenMock(this.inputToken.address);
 
         inputTokenPoolBefore = await this.inputToken.balanceOf(inputTokenAddr);
-        isMain = await this.inputStgToken.isMain.call();
       });
 
       it('normal', async function () {
         // Prep
+        const refundAddress = this.proxy.address;
         const to = this.hStargate.address;
         const fees = await this.layerZeroEndpoint.estimateFees(
           dstChainId,
@@ -1310,12 +1306,13 @@ contract('Stargate', function ([_, user, user2]) {
         );
         const fee = fees[0];
         const data = abi.simpleEncode(
-          'sendTokens(uint16,address,uint256,uint256,uint256)',
+          'sendTokens(uint16,address,address,uint256,uint256,bytes)',
           dstChainId,
           receiver,
+          refundAddress,
           amountIn,
           fee,
-          dstGas
+          adapterParam
         );
 
         // Execute
@@ -1334,22 +1331,19 @@ contract('Stargate', function ([_, user, user2]) {
           ether('0').sub(fee)
         );
 
-        isMain
-          ? expect(
-              await this.inputToken.balanceOf(inputTokenAddr)
-            ).to.be.bignumber.eq(inputTokenPoolBefore.add(amountIn)) // lock token
-          : expect(
-              await this.inputToken.balanceOf(inputTokenAddr)
-            ).to.be.bignumber.eq(inputTokenPoolBefore); // burn token
+        expect(
+          await this.inputToken.balanceOf(inputTokenAddr)
+        ).to.be.bignumber.eq(inputTokenPoolBefore); // burn token
 
         await expectEvent.inTransaction(
           receipt.tx,
           this.inputStgToken,
           'SendToChain',
           {
-            dstChainId: dstChainId.toString(),
-            to: receiver.toString().toLowerCase(),
-            qty: amountIn.toString(),
+            _dstChainId: dstChainId.toString(),
+            _from: this.proxy.address,
+            _toAddress: '0x' + receiverBytes32.toString('hex'),
+            _amount: amountIn.toString(),
           }
         );
 
@@ -1364,6 +1358,7 @@ contract('Stargate', function ([_, user, user2]) {
 
       it('max amount', async function () {
         // Prep
+        const refundAddress = this.proxy.address;
         const to = this.hStargate.address;
         const fees = await this.layerZeroEndpoint.estimateFees(
           dstChainId,
@@ -1374,12 +1369,13 @@ contract('Stargate', function ([_, user, user2]) {
         );
         const fee = fees[0];
         const data = abi.simpleEncode(
-          'sendTokens(uint16,address,uint256,uint256,uint256)',
+          'sendTokens(uint16,address,address,uint256,uint256,bytes)',
           dstChainId,
           receiver,
+          refundAddress,
           MAX_UINT256,
           fee,
-          dstGas
+          adapterParam
         );
 
         // Execute
@@ -1398,22 +1394,19 @@ contract('Stargate', function ([_, user, user2]) {
           ether('0').sub(fee)
         );
 
-        isMain
-          ? expect(
-              await this.inputToken.balanceOf(inputTokenAddr)
-            ).to.be.bignumber.eq(inputTokenPoolBefore.add(amountIn)) // lock token
-          : expect(
-              await this.inputToken.balanceOf(inputTokenAddr)
-            ).to.be.bignumber.eq(inputTokenPoolBefore); // burn token
+        expect(
+          await this.inputToken.balanceOf(inputTokenAddr)
+        ).to.be.bignumber.eq(inputTokenPoolBefore); // burn token
 
         await expectEvent.inTransaction(
           receipt.tx,
           this.inputStgToken,
           'SendToChain',
           {
-            dstChainId: dstChainId.toString(),
-            to: receiver.toString().toLowerCase(),
-            qty: amountIn.toString(),
+            _dstChainId: dstChainId.toString(),
+            _from: this.proxy.address,
+            _toAddress: '0x' + receiverBytes32.toString('hex'),
+            _amount: amountIn.toString(),
           }
         );
 
@@ -1429,10 +1422,11 @@ contract('Stargate', function ([_, user, user2]) {
       it('to a different address', async function () {
         // Prep
         const receiver = user2;
-        const receiverBytes = abi.solidityPack(['address'], [receiver]);
-        const payload = abi.rawEncode(
-          ['bytes', 'uint256'],
-          [receiverBytes, amountIn]
+        const refundAddress = this.proxy.address;
+        const receiverBytes32 = abi.rawEncode(['address'], [receiver]);
+        const payload = abi.solidityPack(
+          ['uint8', 'bytes32', 'uint64'],
+          [ptSend, receiverBytes32, amountIn.div(ld2sdRate)]
         );
         const to = this.hStargate.address;
         const fees = await this.layerZeroEndpoint.estimateFees(
@@ -1444,12 +1438,13 @@ contract('Stargate', function ([_, user, user2]) {
         );
         const fee = fees[0];
         const data = abi.simpleEncode(
-          'sendTokens(uint16,address,uint256,uint256,uint256)',
+          'sendTokens(uint16,address,address,uint256,uint256,bytes)',
           dstChainId,
           receiver,
+          refundAddress,
           amountIn,
           fee,
-          dstGas
+          adapterParam
         );
 
         // Execute
@@ -1468,22 +1463,19 @@ contract('Stargate', function ([_, user, user2]) {
           ether('0').sub(fee)
         );
 
-        isMain
-          ? expect(
-              await this.inputToken.balanceOf(inputTokenAddr)
-            ).to.be.bignumber.eq(inputTokenPoolBefore.add(amountIn)) // lock token
-          : expect(
-              await this.inputToken.balanceOf(inputTokenAddr)
-            ).to.be.bignumber.eq(inputTokenPoolBefore); // burn token
+        expect(
+          await this.inputToken.balanceOf(inputTokenAddr)
+        ).to.be.bignumber.eq(inputTokenPoolBefore); // burn token
 
         await expectEvent.inTransaction(
           receipt.tx,
           this.inputStgToken,
           'SendToChain',
           {
-            dstChainId: dstChainId.toString(),
-            to: receiver.toString().toLowerCase(),
-            qty: amountIn.toString(),
+            _dstChainId: dstChainId.toString(),
+            _from: this.proxy.address,
+            _toAddress: '0x' + receiverBytes32.toString('hex'),
+            _amount: amountIn.toString(),
           }
         );
 
@@ -1498,6 +1490,7 @@ contract('Stargate', function ([_, user, user2]) {
 
       it('refund extra fee', async function () {
         // Prep
+        const refundAddress = this.proxy.address;
         const to = this.hStargate.address;
         const fees = await this.layerZeroEndpoint.estimateFees(
           dstChainId,
@@ -1510,12 +1503,13 @@ contract('Stargate', function ([_, user, user2]) {
         const extraFee = fee.mul(new BN('2'));
         const totalFee = fee.add(extraFee);
         const data = abi.simpleEncode(
-          'sendTokens(uint16,address,uint256,uint256,uint256)',
+          'sendTokens(uint16,address,address,uint256,uint256,bytes)',
           dstChainId,
           receiver,
+          refundAddress,
           amountIn,
           totalFee,
-          dstGas
+          adapterParam
         );
 
         // Execute
@@ -1534,22 +1528,19 @@ contract('Stargate', function ([_, user, user2]) {
           ether('0').sub(totalFee).add(extraFee)
         );
 
-        isMain
-          ? expect(
-              await this.inputToken.balanceOf(inputTokenAddr)
-            ).to.be.bignumber.eq(inputTokenPoolBefore.add(amountIn)) // lock token
-          : expect(
-              await this.inputToken.balanceOf(inputTokenAddr)
-            ).to.be.bignumber.eq(inputTokenPoolBefore); // burn token
+        expect(
+          await this.inputToken.balanceOf(inputTokenAddr)
+        ).to.be.bignumber.eq(inputTokenPoolBefore); // burn token
 
         await expectEvent.inTransaction(
           receipt.tx,
           this.inputStgToken,
           'SendToChain',
           {
-            dstChainId: dstChainId.toString(),
-            to: receiver.toString().toLowerCase(),
-            qty: amountIn.toString(),
+            _dstChainId: dstChainId.toString(),
+            _from: this.proxy.address,
+            _toAddress: '0x' + receiverBytes32.toString('hex'),
+            _amount: amountIn.toString(),
           }
         );
 
@@ -1562,19 +1553,26 @@ contract('Stargate', function ([_, user, user2]) {
         profileGas(receipt);
       });
 
-      // chain
-      it('should revert: to unknown chain', async function () {
+      // adapter param
+      it('should revert: non empty adapter param', async function () {
+        // This test case is only for `useCustomAdapterParams` = false
+        const useCustomAdapterParams =
+          await this.inputStgToken.useCustomAdapterParams.call();
+        if (useCustomAdapterParams) return;
+
         // Prep
-        const dstChainId = STARGATE_UNKNOWN_CHAIN_ID;
+        const adapterParam = abi.solidityPack(['string'], ['0x']);
+        const refundAddress = this.proxy.address;
         const to = this.hStargate.address;
         const fee = ether('1'); // Use fixed fee because of unknown chain id
         const data = abi.simpleEncode(
-          'sendTokens(uint16,address,uint256,uint256,uint256)',
+          'sendTokens(uint16,address,address,uint256,uint256,bytes)',
           dstChainId,
           receiver,
+          refundAddress,
           amountIn,
           fee,
-          dstGas
+          adapterParam
         );
 
         // Execute
@@ -1584,22 +1582,89 @@ contract('Stargate', function ([_, user, user2]) {
             from: user,
             value: value,
           }),
-          '0_HStargate_sendTokens: LayerZero: dstChainId does not exist'
+          '0_HStargate_sendTokens: OFTCore: _adapterParams must be empty.'
+        );
+      });
+
+      it('should revert: zero dstGas', async function () {
+        // This test case is only for `useCustomAdapterParams` = true
+        const useCustomAdapterParams =
+          await this.inputStgToken.useCustomAdapterParams.call();
+        if (!useCustomAdapterParams) return;
+
+        // Prep
+        const version = 1;
+        const dstGas = ether('0');
+        const adapterParam = abi.solidityPack(
+          ['uint16', 'uint256'],
+          [version, dstGas]
+        );
+        const refundAddress = this.proxy.address;
+        const to = this.hStargate.address;
+        const fee = ether('1'); // Use fixed fee because of zero dstGas
+        const data = abi.simpleEncode(
+          'sendTokens(uint16,address,address,uint256,uint256,bytes)',
+          dstChainId,
+          receiver,
+          refundAddress,
+          amountIn,
+          fee,
+          adapterParam
+        );
+
+        // Execute
+        const value = fee;
+        await expectRevert(
+          this.proxy.execMock(to, data, {
+            from: user,
+            value: value,
+          }),
+          '0_HStargate_sendTokens: Relayer: gas too low'
+        );
+      });
+
+      // chain
+      it('should revert: to unknown chain', async function () {
+        // Prep
+        const dstChainId = STARGATE_UNKNOWN_CHAIN_ID;
+        const refundAddress = this.proxy.address;
+        const to = this.hStargate.address;
+        const fee = ether('1'); // Use fixed fee because of unknown chain id
+        const data = abi.simpleEncode(
+          'sendTokens(uint16,address,address,uint256,uint256,bytes)',
+          dstChainId,
+          receiver,
+          refundAddress,
+          amountIn,
+          fee,
+          adapterParam
+        );
+
+        // Execute
+        const value = fee;
+        await expectRevert(
+          this.proxy.execMock(to, data, {
+            from: user,
+            value: value,
+          }),
+          '0_HStargate_sendTokens: LzApp: destination chain is not a trusted source'
         );
       });
 
       // fee
       it('should revert: insufficient fee', async function () {
         // Prep
+        const refundAddress = this.proxy.address;
         const to = this.hStargate.address;
         const fee = ether('0'); // Send zero fee
         const data = abi.simpleEncode(
-          'sendTokens(uint16,address,uint256,uint256,uint256)',
+          'sendTokens(uint16,address,address,uint256,uint256,bytes)',
           dstChainId,
           receiver,
+          refundAddress,
           amountIn,
           fee,
-          dstGas
+          adapterParam
         );
 
         // Execute
@@ -1617,6 +1682,7 @@ contract('Stargate', function ([_, user, user2]) {
       it('should revert: amountIn = 0', async function () {
         // Prep
         const amountIn = ether('0');
+        const refundAddress = this.proxy.address;
         const to = this.hStargate.address;
         const fees = await this.layerZeroEndpoint.estimateFees(
           dstChainId,
@@ -1627,12 +1693,13 @@ contract('Stargate', function ([_, user, user2]) {
         );
         const fee = fees[0];
         const data = abi.simpleEncode(
-          'sendTokens(uint16,address,uint256,uint256,uint256)',
+          'sendTokens(uint16,address,address,uint256,uint256,bytes)',
           dstChainId,
           receiver,
+          refundAddress,
           amountIn,
           fee,
-          dstGas
+          adapterParam
         );
 
         // Execute
@@ -1642,7 +1709,7 @@ contract('Stargate', function ([_, user, user2]) {
             from: user,
             value: value,
           }),
-          '0_HStargate_sendTokens: zero amountIn'
+          '0_HStargate_sendTokens: OFTCore: amount too small'
         );
       });
 
@@ -1650,11 +1717,12 @@ contract('Stargate', function ([_, user, user2]) {
       it('should revert: to zero address', async function () {
         // Prep
         const receiver = constants.ZERO_ADDRESS;
-        const receiverBytes = abi.solidityPack(['address'], [receiver]);
-        const payload = abi.rawEncode(
-          ['bytes', 'uint256'],
-          [receiverBytes, amountIn]
+        const receiverBytes32 = abi.rawEncode(['address'], [receiver]);
+        const payload = abi.solidityPack(
+          ['uint8', 'bytes32', 'uint64'],
+          [ptSend, receiverBytes32, amountIn.div(ld2sdRate)]
         );
+        const refundAddress = this.proxy.address;
         const to = this.hStargate.address;
         const fees = await this.layerZeroEndpoint.estimateFees(
           dstChainId,
@@ -1665,12 +1733,13 @@ contract('Stargate', function ([_, user, user2]) {
         );
         const fee = fees[0];
         const data = abi.simpleEncode(
-          'sendTokens(uint16,address,uint256,uint256,uint256)',
+          'sendTokens(uint16,address,address,uint256,uint256,bytes)',
           dstChainId,
           receiver,
+          refundAddress,
           amountIn,
           fee,
-          dstGas
+          adapterParam
         );
 
         // Execute
@@ -1684,19 +1753,25 @@ contract('Stargate', function ([_, user, user2]) {
         );
       });
 
-      // gas
-      it('should revert: zero dstGas', async function () {
-        // Prep
-        const dstGas = ether('0');
+      it('should revert: refund zero address', async function () {
+        const refundAddress = constants.ZERO_ADDRESS;
         const to = this.hStargate.address;
-        const fee = ether('1'); // Use fixed fee because of zero dstGas
+        const fees = await this.layerZeroEndpoint.estimateFees(
+          dstChainId,
+          this.inputToken.address,
+          payload,
+          false, // pay in zero
+          adapterParam
+        );
+        const fee = fees[0];
         const data = abi.simpleEncode(
-          'sendTokens(uint16,address,uint256,uint256,uint256)',
+          'sendTokens(uint16,address,address,uint256,uint256,bytes)',
           dstChainId,
           receiver,
+          refundAddress,
           amountIn,
           fee,
-          dstGas
+          adapterParam
         );
 
         // Execute
@@ -1706,7 +1781,7 @@ contract('Stargate', function ([_, user, user2]) {
             from: user,
             value: value,
           }),
-          '0_HStargate_sendTokens: Relayer: gas too low'
+          '0_HStargate_sendTokens: refund zero address'
         );
       });
     });